backend/requirements.txt
tensorflow==2.11.0
drf-yasg==1.21.5
Django==4.1.7
djangorestframework==3.13.1
Pillow==9.4.0
django-cors-headers